+++ UICircularProgressRing/UICircularProgressRingDelegate.swift
@@ -45,4 +45,25 @@ public protocol UICircularProgressRingDelegate: class {
      
     */
     func finishedUpdatingProgress(forRing ring: UICircularProgressRingView)
+    
+    /**
+     This method is called whenever the value is updated, this means during animation this method will be called in real time.
+     This can be used to update another label or do some other work, whenever you need the exact current value of the ring
+     during animation.
+     
+     ## Important:
+     
+     This is a very hot method and may be called hundreds of times per second during animations. As such make sure to only
+     do very simple and non-intensive work in this method. Do any work that takes time will considerably slow down your application.
+     
+     - Paramater newValue: The value which the ring has updated to
+     */
+    func didUpdateProgressValue(to newValue: CGFloat)
+}
+
+
+/// Protocol extension to add default conformance
+public extension UICircularProgressRingDelegate {
+    // Adds default conformance with an empty method stub
+    func didUpdateProgressValue(to newValue: CGFloat) { }
 }