+++ run-jazzy.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+jazzy --clean --output docs -a Luis Padron -u https://luispadron.com -m UICircularProgressRing --module-version 1.4.2 -g https://github.com/luispadron/UICircularProgressRing --theme apple